Completed project skeleton and updated docs
